+++ packages/preview-middleware-client/test/unit/adp/controllers/AddFragment.controller.test.ts
@@ -75,7 +75,10 @@ describe('AddFragment', () => {
             const addFragment = new AddFragment(
                 'adp.extension.controllers.AddFragment',
                 overlays as unknown as UI5Element,
-                {} as unknown as RuntimeAuthoring
+                {} as unknown as RuntimeAuthoring,
+                {
+                    title: 'ADP_ADD_FRAGMENT_DIALOG_TITLE'
+                }
             );
 
             const openSpy = jest.fn();
@@ -105,7 +108,10 @@ describe('AddFragment', () => {
             const addFragment = new AddFragment(
                 'adp.extension.controllers.AddFragment',
                 {} as unknown as UI5Element,
-                {} as unknown as RuntimeAuthoring
+                {} as unknown as RuntimeAuthoring,
+                {
+                    title: 'ADP_ADD_FRAGMENT_DIALOG_TITLE'
+                }
             );
 
             const event = {
@@ -163,7 +169,10 @@ describe('AddFragment', () => {
             const addFragment = new AddFragment(
                 'adp.extension.controllers.AddFragment',
                 {} as unknown as UI5Element,
-                {} as unknown as RuntimeAuthoring
+                {} as unknown as RuntimeAuthoring,
+                {
+                    title: 'ADP_ADD_FRAGMENT_DIALOG_TITLE'
+                }
             );
 
             const closeSpy = jest.fn();
@@ -193,7 +202,10 @@ describe('AddFragment', () => {
             const addFragment = new AddFragment(
                 'adp.extension.controllers.AddFragment',
                 {} as unknown as UI5Element,
-                {} as unknown as RuntimeAuthoring
+                {} as unknown as RuntimeAuthoring,
+                {
+                    title: 'ADP_ADD_FRAGMENT_DIALOG_TITLE'
+                }
             );
 
             const valueStateSpy = jest.fn().mockReturnValue({ setValueStateText: jest.fn() });
@@ -223,7 +235,10 @@ describe('AddFragment', () => {
             const addFragment = new AddFragment(
                 'adp.extension.controllers.AddFragment',
                 {} as unknown as UI5Element,
-                {} as unknown as RuntimeAuthoring
+                {} as unknown as RuntimeAuthoring,
+                {
+                    title: 'ADP_ADD_FRAGMENT_DIALOG_TITLE'
+                }
             );
 
             const valueStateSpy = jest.fn().mockReturnValue({ setValueStateText: jest.fn() });
@@ -253,7 +268,10 @@ describe('AddFragment', () => {
             const addFragment = new AddFragment(
                 'adp.extension.controllers.AddFragment',
                 {} as unknown as UI5Element,
-                {} as unknown as RuntimeAuthoring
+                {} as unknown as RuntimeAuthoring,
+                {
+                    title: 'ADP_ADD_FRAGMENT_DIALOG_TITLE'
+                }
             );
 
             const valueStateSpy = jest.fn().mockReturnValue({ setValueStateText: jest.fn() });
@@ -283,7 +301,10 @@ describe('AddFragment', () => {
             const addFragment = new AddFragment(
                 'adp.extension.controllers.AddFragment',
                 {} as unknown as UI5Element,
-                {} as unknown as RuntimeAuthoring
+                {} as unknown as RuntimeAuthoring,
+                {
+                    title: 'ADP_ADD_FRAGMENT_DIALOG_TITLE'
+                }
             );
 
             const valueStateSpy = jest.fn().mockReturnValue({ setValueStateText: jest.fn() });
@@ -313,7 +334,10 @@ describe('AddFragment', () => {
             const addFragment = new AddFragment(
                 'adp.extension.controllers.AddFragment',
                 {} as unknown as UI5Element,
-                {} as unknown as RuntimeAuthoring
+                {} as unknown as RuntimeAuthoring,
+                {
+                    title: 'ADP_ADD_FRAGMENT_DIALOG_TITLE'
+                }
             );
 
             const valueStateSpy = jest.fn().mockReturnValue({ setValueStateText: jest.fn() });
@@ -355,7 +379,10 @@ describe('AddFragment', () => {
             const addFragment = new AddFragment(
                 'adp.extension.controllers.AddFragment',
                 {} as unknown as UI5Element,
-                rtaMock as unknown as RuntimeAuthoring
+                rtaMock as unknown as RuntimeAuthoring,
+                {
+                    title: 'ADP_ADD_FRAGMENT_DIALOG_TITLE'
+                }
             );
 
             const valueStateSpy = jest.fn().mockReturnValue({ setValueStateText: jest.fn() });
@@ -398,7 +425,10 @@ describe('AddFragment', () => {
             const addFragment = new AddFragment(
                 'adp.extension.controllers.AddFragment',
                 {} as unknown as UI5Element,
-                rtaMock as unknown as RuntimeAuthoring
+                rtaMock as unknown as RuntimeAuthoring,
+                {
+                    title: 'ADP_ADD_FRAGMENT_DIALOG_TITLE'
+                }
             );
 
             const valueStateSpy = jest.fn().mockReturnValue({ setValueStateText: jest.fn() });
@@ -445,7 +475,10 @@ describe('AddFragment', () => {
             const addFragment = new AddFragment(
                 'adp.extension.controllers.AddFragment',
                 {} as unknown as UI5Element,
-                rtaMock as unknown as RuntimeAuthoring
+                rtaMock as unknown as RuntimeAuthoring,
+                {
+                    title: 'ADP_ADD_FRAGMENT_DIALOG_TITLE'
+                }
             );
 
             const valueStateSpy = jest.fn().mockReturnValue({ setValueStateText: jest.fn() });
@@ -477,7 +510,10 @@ describe('AddFragment', () => {
             const addFragment = new AddFragment(
                 'adp.extension.controllers.AddFragment',
                 {} as unknown as UI5Element,
-                rtaMock as unknown as RuntimeAuthoring
+                rtaMock as unknown as RuntimeAuthoring,
+                {
+                    title: 'ADP_ADD_FRAGMENT_DIALOG_TITLE'
+                }
             );
 
             const valueStateSpy = jest.fn().mockReturnValue({ setValueStateText: jest.fn() });
@@ -536,7 +572,10 @@ describe('AddFragment', () => {
             const addFragment = new AddFragment(
                 'adp.extension.controllers.AddFragment',
                 overlays as unknown as UI5Element,
-                rtaMock as unknown as RuntimeAuthoring
+                rtaMock as unknown as RuntimeAuthoring,
+                {
+                    title: 'ADP_ADD_FRAGMENT_DIALOG_TITLE'
+                }
             );
 
             const event = {
@@ -604,7 +643,10 @@ describe('AddFragment', () => {
                 'adp.extension.controllers.AddFragment',
                 overlays as unknown as UI5Element,
                 rtaMock as unknown as RuntimeAuthoring,
-                'sections'
+                {
+                    title: 'ADP_ADD_FRAGMENT_DIALOG_TITLE',
+                    aggregation: 'sections'
+                }
             );
 
             const event = {
@@ -689,6 +731,332 @@ describe('AddFragment', () => {
             expect(setContentSpy).toHaveBeenCalledWith({
                 ...dummyContent,
                 templateName: 'OBJECT_PAGE_CUSTOM_SECTION'
+            });
+        });
+
+        test('add header field fragment and a change if targetAggregation is headerContent', async () => {
+            sapMock.ui.version = '1.71.62';
+            const executeSpy = jest.fn();
+            rtaMock.getCommandStack.mockReturnValue({
+                pushAndExecute: executeSpy
+            });
+            rtaMock.getFlexSettings.mockReturnValue({ projectId: 'adp.app' });
+
+            const overlays = {
+                getId: jest.fn().mockReturnValue('some-id')
+            };
+
+            const addFragment = new AddFragment(
+                'adp.extension.controllers.AddFragment',
+                overlays as unknown as UI5Element,
+                rtaMock as unknown as RuntimeAuthoring,
+                {
+                    title: 'QUICK_ACTION_OP_ADD_HEADER_FIELD',
+                    aggregation: 'headerContent'
+                }
+            );
+
+            const event = {
+                getSource: jest.fn().mockReturnValue({
+                    setEnabled: jest.fn()
+                })
+            };
+
+            const testModel = {
+                getProperty: jest
+                    .fn()
+                    .mockReturnValueOnce('Share')
+                    .mockReturnValueOnce('0')
+                    .mockReturnValueOnce('headerContent'),
+                setProperty: jest.fn()
+            } as unknown as JSONModel;
+            addFragment.model = testModel;
+
+            const dummyContent: AddFragmentChangeContentType = {
+                fragmentPath: 'dummyPath',
+                index: 1,
+                targetAggregation: 'headerContent'
+            };
+
+            const setContentSpy = jest.fn();
+            const commandForSpy = jest.fn().mockReturnValue({
+                _oPreparedChange: {
+                    _oDefinition: { moduleName: 'adp/app/changes/fragments/Share.fragment.xml' },
+                    setModuleName: jest.fn()
+                },
+                getPreparedChange: jest.fn().mockReturnValue({
+                    getContent: jest.fn().mockReturnValue(dummyContent),
+                    setContent: setContentSpy
+                })
+            });
+            CommandFactory.getCommandFor = commandForSpy;
+
+            fetchMock.mockResolvedValue({
+                json: jest.fn().mockReturnValue({
+                    id: 'id',
+                    reference: 'reference',
+                    namespace: 'namespace',
+                    layer: 'layer'
+                }),
+                text: jest.fn().mockReturnValue('XML Fragment was created!'),
+                ok: true
+            });
+
+            jest.spyOn(sap.ui, 'getCore').mockReturnValue({
+                byId: jest.fn().mockReturnValue({})
+            } as unknown as Core);
+
+            jest.spyOn(ControlUtils, 'getRuntimeControl').mockReturnValue({
+                getMetadata: jest.fn().mockReturnValue({
+                    getAllAggregations: jest.fn().mockReturnValue({}),
+                    getName: jest.fn().mockReturnValue('sap.uxap.ObjectPageLayout')
+                })
+            } as unknown as ManagedObject);
+
+            addFragment.handleDialogClose = jest.fn();
+
+            await addFragment.setup({
+                setEscapeHandler: jest.fn(),
+                destroy: jest.fn(),
+                setModel: jest.fn(),
+                open: jest.fn(),
+                close: jest.fn()
+            } as unknown as Dialog);
+
+            await addFragment.onCreateBtnPress(event as unknown as Event);
+
+            expect(executeSpy).toHaveBeenCalledWith({
+                _oPreparedChange: {
+                    _oDefinition: {
+                        moduleName: 'adp/app/changes/fragments/Share.fragment.xml'
+                    },
+                    setModuleName: expect.any(Function)
+                },
+                getPreparedChange: expect.any(Function)
+            });
+
+            expect(setContentSpy).toHaveBeenCalledWith({
+                ...dummyContent,
+                templateName: 'OBJECT_PAGE_HEADER_FIELD'
+            });
+        });
+
+        test('add header field fragment and a change if targetAggregation is items', async () => {
+            sapMock.ui.version = '1.71.62';
+            const executeSpy = jest.fn();
+            rtaMock.getCommandStack.mockReturnValue({
+                pushAndExecute: executeSpy
+            });
+            rtaMock.getFlexSettings.mockReturnValue({ projectId: 'adp.app' });
+
+            const overlays = {
+                getId: jest.fn().mockReturnValue('some-id')
+            };
+
+            const addFragment = new AddFragment(
+                'adp.extension.controllers.AddFragment',
+                overlays as unknown as UI5Element,
+                rtaMock as unknown as RuntimeAuthoring,
+                {
+                    title: 'QUICK_ACTION_OP_ADD_HEADER_FIELD',
+                    aggregation: 'items'
+                }
+            );
+
+            const event = {
+                getSource: jest.fn().mockReturnValue({
+                    setEnabled: jest.fn()
+                })
+            };
+
+            const testModel = {
+                getProperty: jest
+                    .fn()
+                    .mockReturnValueOnce('Share')
+                    .mockReturnValueOnce('0')
+                    .mockReturnValueOnce('items'),
+                setProperty: jest.fn()
+            } as unknown as JSONModel;
+            addFragment.model = testModel;
+
+            const dummyContent: AddFragmentChangeContentType = {
+                fragmentPath: 'dummyPath',
+                index: 1,
+                targetAggregation: 'items'
+            };
+
+            const setContentSpy = jest.fn();
+            const commandForSpy = jest.fn().mockReturnValue({
+                _oPreparedChange: {
+                    _oDefinition: { moduleName: 'adp/app/changes/fragments/Share.fragment.xml' },
+                    setModuleName: jest.fn()
+                },
+                getPreparedChange: jest.fn().mockReturnValue({
+                    getContent: jest.fn().mockReturnValue(dummyContent),
+                    setContent: setContentSpy
+                })
+            });
+            CommandFactory.getCommandFor = commandForSpy;
+
+            fetchMock.mockResolvedValue({
+                json: jest.fn().mockReturnValue({
+                    id: 'id',
+                    reference: 'reference',
+                    namespace: 'namespace',
+                    layer: 'layer'
+                }),
+                text: jest.fn().mockReturnValue('XML Fragment was created!'),
+                ok: true
+            });
+
+            jest.spyOn(sap.ui, 'getCore').mockReturnValue({
+                byId: jest.fn().mockReturnValue({})
+            } as unknown as Core);
+
+            jest.spyOn(ControlUtils, 'getRuntimeControl').mockReturnValue({
+                getMetadata: jest.fn().mockReturnValue({
+                    getAllAggregations: jest.fn().mockReturnValue({}),
+                    getName: jest.fn().mockReturnValue('sap.m.FlexBox')
+                }),
+                getParent: jest.fn().mockReturnValue({
+                    getMetadata: jest.fn().mockReturnValue({
+                        getName: jest.fn().mockReturnValue('sap.uxap.ObjectPageDynamicHeaderContent')
+                    })
+                })
+            } as unknown as ManagedObject);
+
+            addFragment.handleDialogClose = jest.fn();
+
+            await addFragment.setup({
+                setEscapeHandler: jest.fn(),
+                destroy: jest.fn(),
+                setModel: jest.fn(),
+                open: jest.fn(),
+                close: jest.fn()
+            } as unknown as Dialog);
+
+            await addFragment.onCreateBtnPress(event as unknown as Event);
+
+            expect(executeSpy).toHaveBeenCalledWith({
+                _oPreparedChange: {
+                    _oDefinition: {
+                        moduleName: 'adp/app/changes/fragments/Share.fragment.xml'
+                    },
+                    setModuleName: expect.any(Function)
+                },
+                getPreparedChange: expect.any(Function)
+            });
+
+            expect(setContentSpy).toHaveBeenCalledWith({
+                ...dummyContent,
+                templateName: 'OBJECT_PAGE_HEADER_FIELD'
+            });
+        });
+
+        test('add header field fragment and a change if targetAggregation is items and not a dynamic header', async () => {
+            sapMock.ui.version = '1.71.62';
+            const executeSpy = jest.fn();
+            rtaMock.getCommandStack.mockReturnValue({
+                pushAndExecute: executeSpy
+            });
+            rtaMock.getFlexSettings.mockReturnValue({ projectId: 'adp.app' });
+
+            const overlays = {
+                getId: jest.fn().mockReturnValue('some-id')
+            };
+
+            const addFragment = new AddFragment(
+                'adp.extension.controllers.AddFragment',
+                overlays as unknown as UI5Element,
+                rtaMock as unknown as RuntimeAuthoring,
+                {
+                    title: 'QUICK_ACTION_OP_ADD_HEADER_FIELD',
+                    aggregation: 'items'
+                }
+            );
+
+            const event = {
+                getSource: jest.fn().mockReturnValue({
+                    setEnabled: jest.fn()
+                })
+            };
+
+            const testModel = {
+                getProperty: jest
+                    .fn()
+                    .mockReturnValueOnce('Share')
+                    .mockReturnValueOnce('0')
+                    .mockReturnValueOnce('items'),
+                setProperty: jest.fn()
+            } as unknown as JSONModel;
+            addFragment.model = testModel;
+
+            const dummyContent: AddFragmentChangeContentType = {
+                fragmentPath: 'dummyPath',
+                index: 1,
+                targetAggregation: 'items'
+            };
+
+            const setContentSpy = jest.fn();
+            const commandForSpy = jest.fn().mockReturnValue({
+                _oPreparedChange: {
+                    _oDefinition: { moduleName: 'adp/app/changes/fragments/Share.fragment.xml' },
+                    setModuleName: jest.fn()
+                },
+                getPreparedChange: jest.fn().mockReturnValue({
+                    getContent: jest.fn().mockReturnValue(dummyContent),
+                    setContent: setContentSpy
+                })
+            });
+            CommandFactory.getCommandFor = commandForSpy;
+
+            fetchMock.mockResolvedValue({
+                json: jest.fn().mockReturnValue({
+                    id: 'id',
+                    reference: 'reference',
+                    namespace: 'namespace',
+                    layer: 'layer'
+                }),
+                text: jest.fn().mockReturnValue('XML Fragment was created!'),
+                ok: true
+            });
+
+            jest.spyOn(sap.ui, 'getCore').mockReturnValue({
+                byId: jest.fn().mockReturnValue({})
+            } as unknown as Core);
+
+            jest.spyOn(ControlUtils, 'getRuntimeControl').mockReturnValue({
+                getMetadata: jest.fn().mockReturnValue({
+                    getAllAggregations: jest.fn().mockReturnValue({}),
+                    getName: jest.fn().mockReturnValue('sap.m.FlexBox')
+                }),
+                getParent: jest.fn().mockReturnValue({
+                    getMetadata: jest.fn().mockReturnValue({
+                        getName: jest.fn().mockReturnValue('No.Dynamic.ObjectPageDynamicHeaderContent')
+                    })
+                })
+            } as unknown as ManagedObject);
+
+            addFragment.handleDialogClose = jest.fn();
+
+            await addFragment.setup({
+                setEscapeHandler: jest.fn(),
+                destroy: jest.fn(),
+                setModel: jest.fn(),
+                open: jest.fn(),
+                close: jest.fn()
+            } as unknown as Dialog);
+
+            await addFragment.onCreateBtnPress(event as unknown as Event);
+
+            expect(executeSpy).toHaveBeenCalledWith({
+                _oPreparedChange: {
+                    _oDefinition: {
+                        moduleName: 'adp/app/changes/fragments/Share.fragment.xml'
+                    },
+                    setModuleName: expect.any(Function)
+                },
+                getPreparedChange: expect.any(Function)
             });
         });
     });